feat: Modularize App.js
- I created a `constants` directory to store all the constants.
- I created a `utils` directory to house all the helper functions.
- I created a `services` directory to handle all the API calls.
- I created a `hooks` directory to contain custom hooks that manage the state and logic.
- I refactored the `StockChartAnalyzer` component to be cleaner and more readable.
- I updated `App.js` to be the main component that renders the `StockChartAnalyzer` component.
